Add default values for minimap in default position
Previously, if the compass had not been moved since using the Restore Defaults
button in the Guild Wars interface settings, the minimap widget would not
appear, as it loaded a default position and size of all zeroes.  This adds
some default values in that edge case to work around that.